transferred task logic for general integrations tasks to new locations
