+++ src/sentry/integrations/bitbucket_server/integration.py
@@ -25,11 +25,11 @@ from sentry.integrations.mixins import RepositoryMixin
 from sentry.integrations.models.integration import Integration
 from sentry.integrations.services.repository import repository_service
 from sentry.integrations.services.repository.model import RpcRepository
+from sentry.integrations.tasks.migrate_repo import migrate_repo
 from sentry.models.identity import Identity
 from sentry.organizations.services.organization import RpcOrganizationSummary
 from sentry.pipeline import PipelineView
 from sentry.shared_integrations.exceptions import ApiError, IntegrationError
-from sentry.tasks.integrations import migrate_repo
 from sentry.web.helpers import render_to_response
 
 from .client import BitbucketServerClient, BitbucketServerSetupClient
