+++ src/sentry/integrations/tasks/create_comment.py
@@ -1,6 +1,10 @@
+from sentry import analytics
+from sentry.integrations.models.external_issue import ExternalIssue
+from sentry.integrations.tasks import should_comment_sync
+from sentry.models.activity import Activity
 from sentry.silo.base import SiloMode, region_silo_function
 from sentry.tasks.base import instrumented_task
-from sentry.tasks.integrations.create_comment import create_comment as old_create_comment
+from sentry.types.activity import ActivityType
 
 
 @region_silo_function
@@ -12,6 +16,29 @@ from sentry.tasks.integrations.create_comment import create_comment as old_creat
     silo_mode=SiloMode.REGION,
 )
 def create_comment(external_issue_id: int, user_id: int, group_note_id: int) -> None:
-    old_create_comment(
-        external_issue_id=external_issue_id, user_id=user_id, group_note_id=group_note_id
+    try:
+        external_issue = ExternalIssue.objects.get(id=external_issue_id)
+    except ExternalIssue.DoesNotExist:
+        return
+
+    installation = external_issue.get_installation()
+
+    if not should_comment_sync(installation, external_issue):
+        return
+
+    try:
+        note = Activity.objects.get(type=ActivityType.NOTE.value, id=group_note_id)
+    except Activity.DoesNotExist:
+        return
+
+    comment = installation.create_comment(external_issue.key, user_id, note)
+    note.data["external_id"] = installation.get_comment_id(comment)
+    note.save()
+    analytics.record(
+        # TODO(lb): this should be changed and/or specified?
+        "integration.issue.comments.synced",
+        provider=installation.model.provider,
+        id=installation.model.id,
+        organization_id=external_issue.organization_id,
+        user_id=user_id,
     )
