+++ src/sentry/integrations/utils/sync.py
@@ -6,12 +6,12 @@ from typing import TYPE_CHECKING
 
 from sentry import features
 from sentry.integrations.services.integration import integration_service
+from sentry.integrations.tasks.sync_assignee_outbound import sync_assignee_outbound
 from sentry.models.group import Group
 from sentry.models.groupassignee import GroupAssignee
 from sentry.models.organization import Organization
 from sentry.models.project import Project
 from sentry.silo.base import region_silo_function
-from sentry.tasks.integrations import sync_assignee_outbound
 from sentry.users.services.user.service import user_service
 
 if TYPE_CHECKING:
