+++ src/sentry/integrations/vsts/tasks/subscription_check.py
@@ -3,11 +3,11 @@ from time import time
 from django.core.exceptions import ObjectDoesNotExist
 
 from sentry.integrations.models.integration import Integration
+from sentry.integrations.tasks import logger
 from sentry.models.apitoken import generate_token
 from sentry.shared_integrations.exceptions import ApiError, ApiUnauthorized
 from sentry.silo.base import SiloMode
 from sentry.tasks.base import instrumented_task, retry
-from sentry.tasks.integrations import logger
 
 
 @instrumented_task(
