+++ src/sentry/models/release.py
@@ -664,6 +664,7 @@ class Release(Model):
         commit_list.sort(key=lambda commit: commit.get("timestamp", 0), reverse=True)
 
         # TODO(dcramer): this function could use some cleanup/refactoring as it's a bit unwieldy
+        from sentry.integrations.tasks.kick_off_status_syncs import kick_off_status_syncs
         from sentry.models.commit import Commit
         from sentry.models.group import Group, GroupStatus
         from sentry.models.grouplink import GroupLink
@@ -673,7 +674,6 @@ class Release(Model):
         from sentry.models.releaseheadcommit import ReleaseHeadCommit
         from sentry.models.repository import Repository
         from sentry.plugins.providers.repository import RepositoryProvider
-        from sentry.tasks.integrations import kick_off_status_syncs
 
         # todo(meredith): implement for IntegrationRepositoryProvider
         commit_list = [
