+++ src/sentry/tasks/integrations/__init__.py
@@ -19,18 +19,10 @@ def should_comment_sync(
 
 
 __all__ = (
-    "create_comment",
-    "kick_off_status_syncs",
     "logger",
     "migrate_opsgenie_plugin",
     "migrate_issues",
-    "migrate_repo",
-    "should_comment_sync",
-    "sync_assignee_outbound",
     "sync_metadata",
-    "sync_status_inbound",
-    "sync_status_outbound",
-    "update_comment",
 )
 
 settings.CELERY_IMPORTS += (
@@ -50,12 +42,6 @@ settings.CELERY_IMPORTS += (
     "sentry.tasks.integrations.vsts.subscription_check",
 )
 
-from .create_comment import create_comment
-from .kick_off_status_syncs_impl import kick_off_status_syncs
+from .migrate_issues import migrate_issues
 from .migrate_opsgenie_plugins import migrate_opsgenie_plugin
-from .migrate_repo import migrate_repo
-from .sync_assignee_outbound_impl import sync_assignee_outbound
 from .sync_metadata import sync_metadata
-from .sync_status_inbound import sync_status_inbound
-from .sync_status_outbound import sync_status_outbound
-from .update_comment import update_comment
