+++ src/sentry/tasks/integrations/kick_off_status_syncs_impl.py
@@ -1,4 +1,6 @@
-from sentry.models.grouplink import GroupLink
+from sentry.integrations.tasks.kick_off_status_syncs import (
+    kick_off_status_syncs as new_kick_off_status_syncs,
+)
 from sentry.silo.base import SiloMode
 from sentry.tasks.base import instrumented_task, retry, track_group_async_operation
 
@@ -13,14 +15,4 @@ from sentry.tasks.base import instrumented_task, retry, track_group_async_operat
 @retry()
 @track_group_async_operation
 def kick_off_status_syncs(project_id: int, group_id: int) -> None:
-    """This is run async to avoid extra queries in the EventManager."""
-    from sentry.tasks.integrations import sync_status_outbound
-
-    external_issue_ids = GroupLink.objects.filter(
-        project_id=project_id, group_id=group_id, linked_type=GroupLink.LinkedType.issue
-    ).values_list("linked_id", flat=True)
-
-    for external_issue_id in external_issue_ids:
-        sync_status_outbound.apply_async(
-            kwargs={"group_id": group_id, "external_issue_id": external_issue_id}
-        )
+    new_kick_off_status_syncs(project_id=project_id, group_id=group_id)
